impl random word with category and close activity after cat selected in CategoryActivity.kt

+++ app/src/main/java/com/pourya/spy_game/service/Api.kt
@@ -13,7 +13,6 @@ interface Api {
     fun getCategories(): Observable<Response<ArrayList<Category>>>
 
     @GET("/v1/spy/categories/{id}")
-    fun getRandomWordWithCategory(@Path("id") categoryId: Long): Observable<Response<Word>>
-
+    fun getRandomWordWithCategory(@Path("id") categoryId: Int): Observable<Response<Word>>
 
 }
